% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.cry-package.R
\docType{package}
\name{rgl.cry-package}
\alias{rgl.cry}
\alias{rgl.cry-package}
\title{rgl.cry: 'cry' and 'rgl' — Applications in Crystallography}
\description{
Visualizing crystal structures and selected area electron diffraction (SAED) patterns. It provides functions cry_demo() and dp_demo() to load a file in 'CIF' (Crystallographic Information Framework) formats and display crystal structures and electron diffraction patterns. The function dp_demo() also performs simple simulation of powder X-ray diffraction (PXRD) patterns, and the results can be saved to a file in the working directory. The package has been tested on several platforms, including Linux on 'Crostini' with a Core™ m3-8100Y Chromebook, I found that even on this low-powered platform, the performance was acceptable. T. Hanashima (2001) \url{https://www2.kek.jp/imss/pf/tools/sasaki/sinram/sinram.html} Todd Helmenstine (2019) \url{https://sciencenotes.org/molecule-atom-colors-cpk-colors/} Wikipedia contributors (2023) \url{https://en.wikipedia.org/w/index.php?title=Atomic_radius&oldid=1179864711}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SaitouToshihide/rgl.cry/}
  \item \url{https://saitoutoshihide.github.io/rgl.cry/}
}

}
\author{
\strong{Maintainer}: Toshihide Saitou \email{saitou.toshihide@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Toshihide Saitou \email{saitou.toshihide@gmail.com} [copyright holder]
}

}
\keyword{internal}
