% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Align crystal and diffraction pattern}
\usage{
align(ax, dev = NULL, verbose = TRUE)
}
\arguments{
\item{ax}{An axis to align}

\item{dev}{RGL device to apply.  Defaults to current device.}

\item{verbose}{logical: Should the report be suppressed?}
}
\value{
No return value, called for side effects.
}
\description{
Align crystal and diffraction pattern and displayed.
}
\details{
There is no z-axis alignment support because the visualization was created
with the analogy of selected area electron diffraction (SAED) on transmission
electron microscope (TEM) which typically have up to two axes.  However you
can rotate around the z-axis by the drag originates near the window edge.
}
\examples{
\donttest{
if (interactive()) {
 align("a")
 align("rb")
 align("1 1 0")
 align("60 -30")
 align(dev = 123, "a")
}
}
}
