% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCIF.R
\name{getCIF}
\alias{getCIF}
\title{Examples of using the cry and rgl packages together.}
\usage{
getCIF(dev = NULL)
}
\arguments{
\item{dev}{The device that is used to extract the lCIF.  The default is
current device.}
}
\value{
A named list, the same as that of 'cry::readCIF.'
}
\description{
Output a lCIF.
}
\examples{
\donttest{
getCIF()
}
}
