% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgl.cry-package.R
\docType{package}
\name{rgl.cry-package}
\alias{rgl.cry}
\alias{rgl.cry-package}
\title{rgl.cry: Use Cases for the Cry and rgl packages: SAED, Crystal Structure Visualization, and PXRD Pattern Simulation}
\description{
Visualizing crystal structures and electron diffraction patterns. It provides functions cry_demo() and dp_demo() to load a file in CIF formats (Crystallographic Information Framework) and display crystal structure and electron diffraction pattern. The function dp_demo() also performs powder diffraction simulation and saves the results to a file in the working directory. The package has been tested on several platforms, including Linux on Crostini with a Core™ m3-8100Y Chromebook, I found that even on this low-powered platform, the performance was acceptable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SaitouToshihide/rgl.cry/}
  \item \url{https://saitoutoshihide.github.io/rgl.cry/}
}

}
\author{
\strong{Maintainer}: Toshihide Saitou \email{saitou.toshihide@gmail.com}

}
\keyword{internal}
